move pkg/apis to api package
