+++ .ci/pipeline_definitions
@@ -32,6 +32,9 @@ gardener-resource-manager:
           preprocess: 'finalize'
         release:
           nextversion: 'bump_minor'
+          git_tags:
+          - ref_template: 'refs/tags/{VERSION}'
+          - ref_template: 'refs/tags/api-{VERSION}'
         slack:
           default_channel: 'internal_scp_workspace'
           channel_cfgs:
