+++ .dockerignore
@@ -2,6 +2,7 @@
 **
 
 # Exclude folders relevant for build
+!api/
 !charts/
 !cmd/
 !pkg/
