+++ api/go.mod
@@ -1,0 +1,8 @@
+module github.com/gardener/gardener-resource-manager/api
+
+go 1.15
+
+require (
+	k8s.io/api v0.18.10
+	k8s.io/apimachinery v0.18.10
+)
