+++ api/resources/register.go
