+++ api/resources/v1alpha1/doc.go
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 // +k8s:deepcopy-gen=package
-// +k8s:conversion-gen=github.com/gardener/gardener-resource-manager/pkg/apis/resources
+// +k8s:conversion-gen=github.com/gardener/gardener-resource-manager/api/resources
 // +k8s:openapi-gen=true
 // +k8s:defaulter-gen=TypeMeta
 
-//go:generate ../../../../hack/update-codegen.sh
+//go:generate ../../../hack/update-codegen.sh
 
 // Package v1alpha1 contains the configuration of the Gardener Resource Manager.
 // +groupName=resources.gardener.cloud
-package v1alpha1 // import "github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1"
+package v1alpha1 // import "github.com/gardener/gardener-resource-manager/api/resources/v1alpha1"
