+++ pkg/apis/resources
@@ -1,0 +1,1 @@
+../../api/resources
