+++ pkg/controller/managedresources/controller.go
@@ -24,8 +24,8 @@ import (
 	"sync"
 	"time"
 
-	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1"
-	resourcesv1alpha1helper "github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1/helper"
+	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/api/resources/v1alpha1"
+	resourcesv1alpha1helper "github.com/gardener/gardener-resource-manager/api/resources/v1alpha1/helper"
 	"github.com/gardener/gardener-resource-manager/pkg/controller/utils"
 
 	hvpav1alpha1 "github.com/gardener/hvpa-controller/api/v1alpha1"
