+++ pkg/controller/managedresources/filter.go
@@ -20,7 +20,7 @@ package managedresources
 import (
 	"strings"
 
-	"github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1"
+	"github.com/gardener/gardener-resource-manager/api/resources/v1alpha1"
 
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/event"
