+++ pkg/controller/managedresources/sort.go
@@ -17,7 +17,7 @@ package managedresources
 import (
 	"sort"
 
-	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1"
+	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/api/resources/v1alpha1"
 )
 
 var _ = sort.Interface(referenceSorter{})
