+++ pkg/mapper/to_mr.go
@@ -17,7 +17,7 @@ package mapper
 import (
 	"context"
 
-	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1"
+	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/api/resources/v1alpha1"
 	"github.com/gardener/gardener-resource-manager/pkg/controller/utils"
 
 	corev1 "k8s.io/api/core/v1"
