+++ pkg/mapper/to_secret_test.go
@@ -15,7 +15,7 @@
 package mapper_test
 
 import (
-	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1"
+	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/api/resources/v1alpha1"
 	"github.com/gardener/gardener-resource-manager/pkg/mapper"
 
 	. "github.com/onsi/ginkgo"
@@ -28,9 +28,7 @@ import (
 )
 
 var _ = Describe("#ManagedResourceToSecretsMapper", func() {
-	var (
-		m handler.Mapper
-	)
+	var m handler.Mapper
 
 	BeforeEach(func() {
 		m = mapper.ManagedResourceToSecretsMapper()
