+++ pkg/predicate/class_changed_test.go
@@ -15,7 +15,7 @@
 package predicate_test
 
 import (
-	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/pkg/apis/resources/v1alpha1"
+	resourcesv1alpha1 "github.com/gardener/gardener-resource-manager/api/resources/v1alpha1"
 	managerpredicate "github.com/gardener/gardener-resource-manager/pkg/predicate"
 
 	. "github.com/onsi/ginkgo"
